[refactor][expr] Separate logical connectives
