+++ scl/cxx/inferenceModule/logic/NegationExpressionNode.hpp
@@ -1,0 +1,29 @@
+/*
+ * This source file is part of an OSTIS project. For the latest info, see http://ostis.net
+ * Distributed under the MIT License
+ * (See accompanying file COPYING.MIT or copy at http://opensource.org/licenses/MIT)
+ */
+
+#pragma once
+
+#include "TemplateExpressionNode.hpp"
+
+using namespace inference;
+
+class NegationExpressionNode : public OperatorLogicExpressionNode
+{
+public:
+  explicit NegationExpressionNode(std::unique_ptr<LogicExpressionNode> op);
+  explicit NegationExpressionNode(ScMemoryContext * context, std::unique_ptr<LogicExpressionNode> op);
+
+  LogicExpressionResult check(ScTemplateParams & params) const override;
+  LogicFormulaResult compute(LogicFormulaResult & result) const override;
+
+  ScAddr getFormulaTemplate() const override
+  {
+    return ScAddr();
+  }
+
+private:
+  ScMemoryContext * context;
+};
